+++ RELEASE_NOTES.md
@@ -75,6 +75,7 @@ Apache Royale Compiler 1.0.0
 - compiler: Fixed intermittent missing `goog.require()` calls when using JS target.
 - compiler: Added support for `mx.core.IDeferredInstance` for default properties with `-children-as-data=false` for JavaScript.
 - compiler: Added support for `mx.core.UIComponentDescriptor` with `-children-as-data=false` for JavaScript.
+- compiler: Improved parsing of `[Exclude]` metadata to make the member name an identifier for tooling.
 - debugger: Added missing isolate ID to SWF load and unload events.
 - debugger: Fixed debugger targeting the current JDK version instead of the intended minimum JDK version.
 - debugger: Fixed localized messages appearing as unprocessed tokens.